Splitter opp requirementResponse og summary of qualifications (#33)
* Lager enkel html for å vise den genererte tabellen
