+++ app/salesGPT/components/employeeCVSummary.module.scss
@@ -7,3 +7,7 @@
   padding: 2rem;
   overflow: auto;
 }
+
+.requirement-response-list{
+  display: none; //TODO toggle on 
+}
